+++ scilla/src/lib.rs
@@ -1,5 +1,6 @@
 mod backend_collector;
 mod call_scilla_server;
 mod proto;
+mod types;
 pub mod scilla_server_run;
 pub mod scilla_tcp_server;
